pyproject.toml
[tool.poetry]
name = "fastapi-frame"
version = "0.1.0"
description = ""
authors = ["loong_zhou <zhoulf1006@gmail.com>"]
readme = "README.md"
packages = [{ include = "fastapi_frame" }]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.111.0"
pydantic = "^2.8.2"
sqlmodel = "^0.0.21"
pydantic-settings = "^2.3.4"
sqlalchemy = "^2.0.31"
asyncpg = "^0.29.0"
greenlet = "^3.0.3"
alembic = "^1.13.2"
uvicorn = "^0.30.1"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
tenacity = "^8.5.0"
fastapi-versionizer = "^4.0.1"
fastapi-pagination = "^0.12.26"
redis = "^5.0.7"
botocore = "^1.34.144"
cachetools = "^5.4.0"
prometheus-fastapi-instrumentator = "^7.0.0"
opentelemetry-api = "^1.25.0"
opentelemetry-sdk = "^1.25.0"
opentelemetry-instrumentation-fastapi = "^0.46b0"
opentelemetry-exporter-otlp-proto-http = "^1.25.0"
opentelemetry-exporter-otlp-proto-grpc = "^1.25.0"
urllib3 = ">=1.25.4,<1.27"
msal = "^1.30.0"
alembic-postgresql-enum = "^1.3.0"
pyjwt = "^2.9.0"
asynch = "^0.2.4"
clickhouse-driver = "^0.2.9"
azure-kusto-data = "^4.5.1"
pandas = "^2.2.2"
aiomysql = "^0.2.0"
slack-sdk = "^3.33.1"
pymsteams = "^0.2.3"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"
ruff = "^0.5.1"
mypy = "^1.10.1"
pytest = "^8.3.2"
pytest-asyncio = "^0.23.8"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff] # https://docs.astral.sh/ruff/settings/#top-level
line-length = 100
exclude = ["**/migrations/", "**/manage.py"]

[tool.ruff.lint] # https://docs.astral.sh/ruff/settings/#lint
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "UP", # pyupgrade
]
ignore = [
    "E402", # module level import not at top of file
]

[tool.ruff.format] # https://docs.astral.sh/ruff/settings/#format
quote-style = "double"
